# Phase 2: Memory & Scheduling Optimization - Complete Implementation